requirements.txt
# Parakeet-Writer Requirements
# Use with: pip install -r requirements.txt
# NOTE: Install PyTorch first with CUDA support using the command below:
# pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu121

# Core ML/AI Dependencies
omegaconf>=2.3.0
hydra-core>=1.3.2
nemo-toolkit[asr]>=2.3.1

# Audio Processing
sounddevice>=0.5.0
scipy>=1.15.0
librosa>=0.10.1
soundfile>=0.13.0
audioplayer>=0.6
resampy>=0.4.3
pydub>=0.25.1

# GUI and Desktop Integration  
PyQt5>=5.15.11
pynput>=1.8.0

# Voice Activity Detection
webrtcvad>=2.0.10

# Environment and Configuration
python-dotenv>=1.0.0

# API Integration
openai>=1.86.0

# GPU Optimization (Recommended)
cuda-python>=12.9.0

# General Utilities
numpy>=1.26.0,<2.0.0
tqdm>=4.67.0
requests>=2.32.0
setuptools>=80.0.0

# Optional: Additional performance packages
# accelerate>=1.7.0
# transformers>=4.51.0 